blogilistan laajennus 4.21

+++ osa4/blogilista/controllers/blogs.js
@@ -61,8 +61,15 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const token = getTokenFrom(request)
+  const decodedToken = jwt.verify(token, process.env.SECRET)
+  const userDel = await User.findById(decodedToken.id)
+  const blogDel = await Blog.findById(request.params.id)
+  if (blogDel.user.toString() === userDel.id.toString()){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else {response.status(400).end()}
 })
 
 module.exports = blogsRouter
